Platform/ARM: Reorganize Lcd Graphics Output
This corresponds to the recently submitted edk2 change "ArmPlatformPkg:
Reorganize Lcd Graphics Output".

This change enables building of a common LcdGraphicsOutputDxe,
replacing PL111LcdGraphicsOutputDxe and HdLcdGraphicsOutputDxe.

One of the different hardware implementations (PL111Lcd, HdLcd, and
MaliDp) is included as a LcdHwLib library.

NOTE: The FVP changes include framework for HdLcd and MaliDp builds.
These are not part of the supported FVP model, but may be added as
customisations.  Because the parameters that might be used are not
known, some PCD settings are commented out.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Girish Pathak <girish.pathak@arm.com>
Signed-off-by: Evan Lloyd <evan.lloyd@arm.com>