ARM/JunoPkg: Mapping Non-Trused SRAM as device memory
This fix changes the cache attribute of Non-Trusted SRAM on the Juno
platform to device memory. This change is required to avoid coherency
problems as Non-Trusted SRAM is used as a shared memory between the
application processor and the SCP for communication. This change is a
prerequisite for upcoming SCMI driver for the Juno platform.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Girish Pathak <girish.pathak@arm.com>